Safari-609.1.11.1/LayoutTests/fast/dom/Window/window-open-self-disallow-close-expected.txt

CONSOLE MESSAGE: line 8: Can't close the window since it was not opened by JavaScript

PASS Calling window.open with windowName '_self' should not make the window closeable by script 
